Renamed main class to CafeCtrl
