+++ build.gradle
@@ -29,11 +29,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("seedu.cafectrl.CafeCtrl")
 }
 
 shadowJar {
-    archiveBaseName.set("duke")
+    archiveBaseName.set("cafectrl")
     archiveClassifier.set("")
 }
