+++ src/main/java/seedu/cafectrl/command/IncorrectCommand.java
@@ -1,0 +1,20 @@
+package seedu.cafectrl.command;
+
+import seedu.cafectrl.data.Menu;
+import seedu.cafectrl.ui.Ui;
+
+/**
+ * Represents an incorrect command. Upon execution, produces some feedback to the user.
+ */
+public class IncorrectCommand extends Command{
+    public final String feedbackToUser;
+
+    public IncorrectCommand(String feedbackToUser) {
+        this.feedbackToUser = feedbackToUser;
+    }
+
+    @Override
+    public void execute(Menu menu, Ui ui) {
+        //ui.showToUser(feedbackToUser);
+    }
+}
