--- src/main/java/seedu/duke/command/IncorrectCommand.java
@@ -1,20 +1,0 @@
-package seedu.duke.command;
-
-import seedu.duke.data.Menu;
-import seedu.duke.ui.Ui;
-
-/**
- * Represents an incorrect command. Upon execution, produces some feedback to the user.
- */
-public class IncorrectCommand extends Command{
-    public final String feedbackToUser;
-
-    public IncorrectCommand(String feedbackToUser) {
-        this.feedbackToUser = feedbackToUser;
-    }
-
-    @Override
-    public void execute(Menu menu, Ui ui) {
-        //ui.showToUser(feedbackToUser);
-    }
-}
