--- src/main/java/seedu/duke/command/ListMenuCommand.java
@@ -1,31 +1,0 @@
-package seedu.duke.command;
-
-import seedu.duke.data.Menu;
-import seedu.duke.ui.Ui;
-import seedu.duke.ui.UserOutput;
-
-import java.text.DecimalFormat;
-
-/**
- * Lists all dishes in the menu to the user.
- */
-public class ListMenuCommand extends Command {
-    public static final String COMMAND_WORD = "list";
-    private static final DecimalFormat dollarValue = new DecimalFormat("0.00");
-    /**
-     * Iterates through the menu arraylist, outputting the dish name and dish price.
-     *
-     * @param menu ArrayList of Dishes
-     * @param ui Handles the interactions with user
-     */
-    @Override
-    public void execute(Menu menu, Ui ui) {
-        ui.showToUser(UserOutput.LIST_MESSAGE.message);
-        for(int i =0; i < menu.getSize(); i++) {
-            String indexNum = String.valueOf(i+1);
-            String dishName = menu.getDish(i).getName();
-            String dishPrice = dollarValue.format(menu.getDish(i).getPrice());
-            ui.showToUser(indexNum + "." + dishName + " $" + dishPrice);
-        }
-    };
-}
