+++ src/test/java/seedu/cafectrl/DukeTest.java
@@ -1,0 +1,12 @@
+package seedu.cafectrl;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+class DukeTest {
+    @Test
+    public void sampleTest() {
+        assertTrue(true);
+    }
+}
