+++ src/test/java/seedu/cafectrl/UiTest.java
@@ -1,0 +1,8 @@
+package seedu.cafectrl;
+
+
+/**
+ * Junit test for Ui.java
+ */
+public class UiTest {
+}
