+++ src/test/java/seedu/cafectrl/command/AddDishCommandTest.java
@@ -1,0 +1,5 @@
+package seedu.cafectrl.command;
+
+class AddDishCommandTest {
+
+}
