+++ src/test/java/seedu/cafectrl/command/CommandTest.java
@@ -1,0 +1,5 @@
+package seedu.cafectrl.command;
+
+class CommandTest {
+
+}
