+++ src/test/java/seedu/cafectrl/command/DeleteDishCommandTest.java
@@ -1,0 +1,5 @@
+package seedu.cafectrl.command;
+
+class DeleteDishCommandTest {
+
+}
