+++ src/test/java/seedu/cafectrl/data/MenuTest.java
@@ -1,0 +1,7 @@
+package seedu.cafectrl.data;
+
+//import static org.junit.jupiter.api.Assertions.*;
+
+class MenuTest {
+
+}
