+++ src/test/java/seedu/cafectrl/data/dish/DishTest.java
@@ -1,0 +1,7 @@
+package seedu.cafectrl.data.dish;
+
+//import static org.junit.jupiter.api.Assertions.*;
+
+class DishTest {
+
+}
