+++ src/test/java/seedu/cafectrl/parser/ParserTest.java
@@ -1,0 +1,8 @@
+package seedu.cafectrl.parser;
+
+/**
+ * Junit test for Parser.java
+ */
+class ParserTest {
+
+}
