--- src/test/java/seedu/duke/UiTest.java
@@ -1,8 +1,0 @@
-package seedu.duke;
-
-
-/**
- * Junit test for Ui.java
- */
-public class UiTest {
-}
