--- src/test/java/seedu/duke/command/ListIngredientCommandTest.java
@@ -1,5 +1,0 @@
-package seedu.duke.command;
-
-class ListIngredientCommandTest {
-
-}
