--- src/test/java/seedu/duke/command/ListMenuCommandTest.java
@@ -1,21 +1,0 @@
-package seedu.duke.command;
-
-import org.junit.jupiter.api.Test;
-import seedu.duke.data.Menu;
-import seedu.duke.data.dish.Dish;
-import seedu.duke.ui.Ui;
-
-import java.util.ArrayList;
-
-class ListMenuCommandTest {
-    @Test
-    public static void execute_expectTwoDishes() {
-        ArrayList<Dish> menuItems = new ArrayList<>();
-        menuItems.add(new Dish("Chicken Rice", 2.50F));
-        menuItems.add(new Dish("Chicken Curry", 4.30F));
-        Menu menu = new Menu(menuItems);
-
-        Command listMenuCommand = new ListMenuCommand();
-        listMenuCommand.execute(menu, new Ui());
-    }
-}
