--- src/test/java/seedu/duke/data/MenuTest.java
@@ -1,7 +1,0 @@
-package seedu.duke.data;
-
-//import static org.junit.jupiter.api.Assertions.*;
-
-class MenuTest {
-
-}
