--- src/test/java/seedu/duke/data/dish/IngredientTest.java
@@ -1,7 +1,0 @@
-package seedu.duke.data.dish;
-
-//import static org.junit.jupiter.api.Assertions.*;
-
-class IngredientTest {
-
-}
