--- src/test/java/seedu/duke/parser/ParserTest.java
@@ -1,8 +1,0 @@
-package seedu.duke.parser;
-
-/**
- * Junit test for Parser.java
- */
-class ParserTest {
-
-}
